Create separate settings package under UI
